Moved dotnotation to backend-utils
